# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nova Kudos",
        "description": "Demo Nova Kudos Api",
        "version": "1.0"
    },
    "paths": {
...
        "/api/app/users/all": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "get all users",
                "description": "get all users",
                "operationId": "get all users",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "/api/app/users/all"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "validation error"
                    },
                    "422": {
                        "description": "error"
                    },
                    "500": {
                        "description": "server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/users/store": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "store new user",
                "description": "store new user",
                "operationId": "store new user",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "mobile"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "yasin"
                                    },
                                    "last_name": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "baghban"
                                    },
                                    "mobile": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "091010101010"
                                    },
                                    "avatar": {
                                        "type": "file",
                                        "format": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "401": {
                        "description": "validation error"
                    },
                    "400": {
                        "description": "error"
                    },
                    "500": {
                        "description": "server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/users/update/{user}": {
            "put": {
                "tags": [
                    "users"
                ],
                "summary": "update user",
                "description": "update user",
                "operationId": "update user",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "first_name"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "yasin"
                                    },
                                    "last_name": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "baghban"
                                    },
                                    "mobile": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "09354068701"
                                    },
                                    "avatar": {
                                        "type": "text",
                                        "format": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "401": {
                        "description": "validation error"
                    },
                    "400": {
                        "description": "error"
                    },
                    "500": {
                        "description": "server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/users/delete/{user}": {
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "delete user",
                "description": "delete user",
                "operationId": "delete user",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}